LPS-77989 This is a data object, no need for interface
