+++ modules/apps/static/portal-osgi-web/portal-osgi-web-servlet-context-helper/src/main/java/com/liferay/portal/osgi/web/servlet/context/helper/order/Order.java
@@ -14,34 +14,94 @@
 
 package com.liferay.portal.osgi.web.servlet.context.helper.order;
 
+import com.liferay.portal.kernel.util.ArrayUtil;
+
+import java.util.Arrays;
 import java.util.EnumMap;
 
 /**
  * @author Vernon Singleton
  * @author Juan Gonzalez
  */
-public interface Order {
+public class Order {
 
 	public static final String OTHERS = Order.class.getName() + "#OTHERS";
 
-	public EnumMap<Path, String[]> getRoutes();
+	public Order() {
+		_routes = new EnumMap<>(Path.class);
 
-	public boolean isAfter(String name);
+		_routes.put(Path.BEFORE, new String[0]);
+		_routes.put(Path.AFTER, new String[0]);
+	}
 
-	public boolean isAfterOthers();
+	public EnumMap<Path, String[]> getRoutes() {
+		return _routes;
+	}
 
-	public boolean isBefore(String name);
+	public boolean isAfter(String name) {
+		if (Arrays.binarySearch(_routes.get(Path.AFTER), name) >= 0) {
+			return true;
+		}
 
-	public boolean isBeforeOthers();
+		return false;
+	}
 
-	public boolean isOrdered();
+	public boolean isAfterOthers() {
+		boolean value = false;
 
-	public void setRoutes(EnumMap<Path, String[]> routes);
+		if (_routes.get(Path.AFTER) != null) {
+			if (Arrays.binarySearch(_routes.get(Path.AFTER), Order.OTHERS) >=
+					0) {
+
+				value = true;
+			}
+		}
+
+		return value;
+	}
+
+	public boolean isBefore(String name) {
+		if (Arrays.binarySearch(_routes.get(Path.BEFORE), name) >= 0) {
+			return true;
+		}
+
+		return false;
+	}
+
+	public boolean isBeforeOthers() {
+		boolean value = false;
+
+		if (_routes.get(Path.BEFORE) != null) {
+			if (Arrays.binarySearch(_routes.get(Path.BEFORE), Order.OTHERS) >=
+					0) {
+
+				value = true;
+			}
+		}
+
+		return value;
+	}
+
+	public boolean isOrdered() {
+		if (ArrayUtil.isNotEmpty(_routes.get(Path.AFTER)) ||
+			ArrayUtil.isNotEmpty(_routes.get(Path.BEFORE))) {
+
+			return true;
+		}
+
+		return false;
+	}
+
+	public void setRoutes(EnumMap<Path, String[]> routes) {
+		_routes = routes;
+	}
 
 	public enum Path {
 
 		BEFORE, AFTER
 
 	}
+
+	private EnumMap<Path, String[]> _routes;
 
 }
